test(app): further densify naval panel family for #4224 AC7
Compress combine-outcome scenario tables, centralize outcome pumping,
refactor mockup-fidelity onto shared game/pump helpers, and trim support
home-transfer wrappers. Part shards 1227→1027 (−16%); family 3217→3103.

Refs #4224
