+++ app/test/naval_units_panel_mockup_fidelity_test.dart
@@ -3,8 +3,6 @@
 // `SPEC/ui/naval-units-panel.md`. Each `group` corresponds to one R-item so
 // regressions point straight at the spec line they broke.
 
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart' show homeFleetIdFor;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
@@ -18,97 +16,7 @@ import 'package:colonizethis_app/features/game/widgets/units/shared/units_entity
 import 'package:colonizethis_app/features/game/widgets/units/shared/units_entity_card.dart';
 import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
 
-import 'app_shell_harness.dart';
 import 'naval_units_panel_test_support.dart';
-
-const _humanId = 'gp_naval_fidelity';
-const _capitalLocalId = 'cap1';
-const _capitalProvinceId = 'oldWorld|$_capitalLocalId';
-const _portLocalId = 'port1';
-const _portProvinceId = 'oldWorld|$_portLocalId';
-const _localSeaZoneId = 'zone_alpha';
-const _zonePrefixedId = 'oldWorld|$_localSeaZoneId';
-
-/// Deterministic mockup-fidelity scenario via shared OW fleets factory
-/// (Refs #4021 — no private `Game(` in naval units-panel suites).
-Game _buildFidelityGame() {
-  final homeId = homeFleetIdFor(_humanId);
-  return buildNavalPanelOwFleetsGame(
-    gameId: 'naval-fidelity',
-    humanId: _humanId,
-    displayName: 'Fidelity Tester',
-    capitalProvinceId: _capitalProvinceId,
-    oldWorldProvinces: const [
-      Province(
-        id: _capitalLocalId,
-        regionId: 'oldWorld',
-        ownerId: _humanId,
-        displayName: 'London',
-      ),
-      Province(
-        id: _portLocalId,
-        regionId: 'oldWorld',
-        ownerId: _humanId,
-        displayName: 'Portsmouth',
-      ),
-    ],
-    fleets: [
-      // Home Fleet (in port at capital) — drives R26 HOME chip and R29
-      // Home-Fleet cargo line.
-      Fleet(
-        id: homeId,
-        ownerId: _humanId,
-        regionId: 'oldWorld',
-        inPortAtProvinceId: _capitalProvinceId,
-        ships: const [
-          ShipInstance(id: 'h1', typeId: 'carrack'),
-          ShipInstance(id: 'h2', typeId: 'frigate'),
-        ],
-      ),
-      // In-port sea-going fleet — drives R25 dense pills, R27 locate
-      // alignment, R28 `(in port)` qualifier.
-      Fleet(
-        id: 'channel_fleet',
-        ownerId: _humanId,
-        regionId: 'oldWorld',
-        inPortAtProvinceId: _portProvinceId,
-        ships: const [
-          ShipInstance(id: 'p1', typeId: 'frigate'),
-          ShipInstance(id: 'p2', typeId: 'frigate'),
-        ],
-      ),
-      // At-sea fleet — drives R28 `(at sea)` qualifier.
-      Fleet(
-        id: 'atlantic_fleet',
-        ownerId: _humanId,
-        regionId: 'oldWorld',
-        seaZoneId: _localSeaZoneId,
-        ships: const [ShipInstance(id: 's1', typeId: 'galleon')],
-      ),
-    ],
-    seaZoneDisplayNameById: const {_zonePrefixedId: 'Bay of Biscay'},
-    tileKeysByProvince: const {
-      _capitalProvinceId: ['oldWorld|$_capitalLocalId|0|0'],
-      _portProvinceId: ['oldWorld|$_portLocalId|0|0'],
-    },
-  );
-}
-
-Widget _hostPanel(Game game) {
-  return buildAppShell(
-    child: Scaffold(
-      body: ConstrainedBox(
-        constraints: const BoxConstraints(maxWidth: 480, maxHeight: 720),
-        child: NavalUnitsPanel(
-          game: game,
-          humanPlayerId: _humanId,
-          bus: AppEventBus.create(),
-          topology: const MapTopology(),
-        ),
-      ),
-    ),
-  );
-}
 
 void main() {
   suppressLogsForTests();
@@ -116,19 +24,17 @@ void main() {
 
   late Game game;
   setUpAll(() {
-    // Build the deterministic fidelity scenario directly. Refs #3656: the panel
-    // renders entirely from this hand-built fixture, so the prior ~11s
-    // `getDebugInitGameResult()` warm call (which generated a full map this
-    // suite never reads) is removed.
-    game = _buildFidelityGame();
+    game = buildNavalPanelMockupFidelityGame();
   });
+
+  Future<void> pumpFidelity(WidgetTester tester) =>
+      pumpNavalMockupFidelityPanel(tester, game: game);
 
   group('R25 — compact inline action pills on one row', () {
     testWidgets(
       'Non-home fleet renders Move + Split + Locate on a single dense row',
       (WidgetTester tester) async {
-        await tester.pumpWidget(_hostPanel(game));
-        await tester.pumpAndSettle();
+        await pumpFidelity(tester);
 
         final channelTile = find.widgetWithText(
           ExpansionTile,
@@ -142,15 +48,8 @@ void main() {
             matching: find.byType(UnitsEntityActionRow),
           ),
         );
-        expect(
-          actionRow.dense,
-          isTrue,
-          reason: 'Naval fleet rows must use the dense pill footprint (R25)',
-        );
+        expect(actionRow.dense, isTrue);
 
-        // Move + Split render as mockup compact-pill CtActionTextButtons and
-        // Locate as the circular CtCircularLocateButton (issue #3514 owner
-        // decision #6); the row must mount no CtNinePatchButton action chrome.
         final pillButtons = find.descendant(
           of: channelTile,
           matching: find.byType(CtActionTextButton),
@@ -167,20 +66,14 @@ void main() {
             matching: find.byType(CtNinePatchButton),
           ),
           findsNothing,
-          reason: 'Naval fleet row actions must use pills, not nine-patch (R25)',
         );
 
-        // All three controls render on a single horizontal line (no wrap).
         final yCenters = <double>{
           tester.getCenter(pillButtons.first).dy,
           tester.getCenter(pillButtons.last).dy,
           tester.getCenter(locateButton).dy,
         };
-        expect(
-          yCenters.length,
-          1,
-          reason: 'Move/Split/Locate must share one row (R25)',
-        );
+        expect(yCenters.length, 1);
       },
     );
   });
@@ -189,16 +82,13 @@ void main() {
     testWidgets(
       'HOME chip renders next to Home Fleet name and not on regular fleets',
       (WidgetTester tester) async {
-        await tester.pumpWidget(_hostPanel(game));
-        await tester.pumpAndSettle();
+        await pumpFidelity(tester);
 
         final homeTile = find.widgetWithText(ExpansionTile, 'Home Fleet');
         expect(homeTile, findsOneWidget);
-
         expect(
           find.descendant(of: homeTile, matching: find.text('HOME')),
           findsOneWidget,
-          reason: 'Home Fleet row must render the uppercase HOME chip (R26)',
         );
 
         final channelTile = find.widgetWithText(
@@ -208,7 +98,6 @@ void main() {
         expect(
           find.descendant(of: channelTile, matching: find.text('HOME')),
           findsNothing,
-          reason: 'Regular fleet rows must NOT render the HOME chip (R26)',
         );
       },
     );
@@ -220,8 +109,7 @@ void main() {
       testWidgets(
         'Locate is rightmost on channel fleet; Home Fleet is Split + Locate',
         (WidgetTester tester) async {
-          await tester.pumpWidget(_hostPanel(game));
-          await tester.pumpAndSettle();
+          await pumpFidelity(tester);
 
           final channelTile = find.widgetWithText(
             ExpansionTile,
@@ -260,7 +148,6 @@ void main() {
               matching: find.byType(Text),
             ),
             findsNothing,
-            reason: 'Locate must be icon-only (R27 mockup `.locate-btn`)',
           );
 
           final homeTile = find.widgetWithText(ExpansionTile, 'Home Fleet');
@@ -288,23 +175,10 @@ void main() {
     testWidgets('In-port and at-sea fleets append localised qualifiers', (
       WidgetTester tester,
     ) async {
-      await tester.pumpWidget(_hostPanel(game));
-      await tester.pumpAndSettle();
+      await pumpFidelity(tester);
 
-      expect(
-        find.text('Old World — Portsmouth (in port)'),
-        findsOneWidget,
-        reason:
-            'R28: in-port fleet location must end with the localised '
-            '`(in port)` qualifier',
-      );
-      expect(
-        find.text('Old World — Bay of Biscay (at sea)'),
-        findsOneWidget,
-        reason:
-            'R28: at-sea fleet location must end with the localised '
-            '`(at sea)` qualifier',
-      );
+      expect(find.text('Old World — Portsmouth (in port)'), findsOneWidget);
+      expect(find.text('Old World — Bay of Biscay (at sea)'), findsOneWidget);
     });
   });
 
@@ -312,8 +186,7 @@ void main() {
     testWidgets(
       'Home Fleet expanded: Table rows, cargo, summary; non-home omits cargo',
       (WidgetTester tester) async {
-        await tester.pumpWidget(_hostPanel(game));
-        await tester.pumpAndSettle();
+        await pumpFidelity(tester);
 
         final homeTile = find.widgetWithText(ExpansionTile, 'Home Fleet');
         await tester.tap(homeTile);
@@ -323,35 +196,17 @@ void main() {
           of: homeTile,
           matching: find.byType(Table),
         );
-        expect(
-          tableFinder,
-          findsOneWidget,
-          reason: 'R29: expanded view must render a Table widget',
-        );
-        // Two ship types in the Home Fleet (carrack + frigate).
+        expect(tableFinder, findsOneWidget);
         expect(tester.widget<Table>(tableFinder).children.length, 2);
-
         expect(
           find.text('Total ships: 2 · Warships: 1 · Merchants: 1'),
           findsOneWidget,
-          reason:
-              'R29: composition summary must be a single Text widget, not '
-              'three separate ListTiles',
         );
         expect(find.text('1 warships'), findsNothing);
         expect(find.text('1 merchants'), findsNothing);
         expect(find.text('Total ships: 2'), findsNothing);
-
-        expect(
-          find.textContaining('Cargo capacity:'),
-          findsOneWidget,
-          reason: 'R29: Home Fleet expanded view must render the cargo line',
-        );
-        expect(
-          find.textContaining('holds'),
-          findsOneWidget,
-          reason: 'R29: cargo line uses the `holds` localisation key',
-        );
+        expect(find.textContaining('Cargo capacity:'), findsOneWidget);
+        expect(find.textContaining('holds'), findsOneWidget);
         expect(find.textContaining('Strength:'), findsOneWidget);
 
         final channelTile = find.widgetWithText(
@@ -367,9 +222,6 @@ void main() {
             matching: find.textContaining('Cargo capacity'),
           ),
           findsNothing,
-          reason:
-              'R29: per mockup, non-home fleets do not render a cargo line '
-              'in the expanded view',
         );
         expect(
           find.descendant(
@@ -386,36 +238,20 @@ void main() {
     testWidgets('FleetExpansionTile is the canonical naval row widget', (
       WidgetTester tester,
     ) async {
-      await tester.pumpWidget(_hostPanel(game));
-      await tester.pumpAndSettle();
-      // 1 Home Fleet + 2 regular fleets in the deterministic fixture.
+      await pumpFidelity(tester);
       expect(find.byType(FleetExpansionTile), findsNWidgets(3));
     });
   });
 
-  // Issue #3514 AC-6 (naval card migration): fleet rows render the shared
-  // UnitsEntityCard mockup `.fleet-row` bordered gradient card chrome (the
-  // same migration delivered for military army rows), not the bare Material
-  // ExpansionTile chrome. The deferred-from-#3532 dense-action overflow under
-  // the card chrome is also pinned (no RenderFlex exception at the default
-  // 480-dp host).
   group('Naval fleet card chrome (issue #3514 AC-6)', () {
     testWidgets(
       'Each fleet row is rendered inside a UnitsEntityCard with its dense '
       'action row hosted chrome-less (no double border) and no overflow',
       (WidgetTester tester) async {
-        await tester.pumpWidget(_hostPanel(game));
-        await tester.pumpAndSettle();
-
-        // No RenderFlex overflow exception escapes when the dense Move /
-        // Split / Locate cluster lays out under the card chrome (the gap
-        // PR #3532 deferred for naval).
+        await pumpFidelity(tester);
         expect(tester.takeException(), isNull);
-
-        // One UnitsEntityCard per fleet row (Home + 2 regular fleets).
         expect(find.byType(UnitsEntityCard), findsNWidgets(3));
 
-        // Each FleetExpansionTile hosts its card chrome.
         for (final card in <Finder>[
           find.descendant(
             of: find.byType(FleetExpansionTile).at(0),
@@ -433,25 +269,13 @@ void main() {
           expect(card, findsOneWidget);
         }
 
-        // The dense title action row is hosted chrome-less so the card border
-        // is not double-painted (issue #3514 AC-6 — `chrome: false`).
         final actionRows = tester.widgetList<UnitsEntityActionRow>(
           find.byType(UnitsEntityActionRow),
         );
         expect(actionRows, isNotEmpty);
         for (final row in actionRows) {
-          expect(
-            row.chrome,
-            isFalse,
-            reason:
-                'Naval fleet action rows must be hosted chrome-less inside '
-                'the UnitsEntityCard (issue #3514 AC-6)',
-          );
-          expect(
-            row.dense,
-            isTrue,
-            reason: 'Naval fleet action rows stay dense (R25)',
-          );
+          expect(row.chrome, isFalse);
+          expect(row.dense, isTrue);
         }
       },
     );
